requirements.txt
streamlit
crewai
langchain_groq
crewai_tools
python-dotenv
pandas
ipython
opencv-python
pytesseract
embedchain